zip folder after downloading. added print statements

+++ eapbl-project.py
@@ -3,15 +3,16 @@ from bs4 import BeautifulSoup
 import os
 import csv
 import zipfile
+import shutil
 
 '''
 Program to download images from the endangered archives collection from http://eap.bl.uk/database/collections.a4d
 '''
 
-# list of project urls to download
-urls = ['http://eap.bl.uk/database/results.a4d?projID=EAP183',
-        'http://eap.bl.uk/database/results.a4d?projID=EAP314',
-        'http://eap.bl.uk/database/results.a4d?projID=EAP458'
+# list of project urls to downloads
+urls = [#'http://eap.bl.uk/database/results.a4d?projID=EAP183',
+        # 'http://eap.bl.uk/database/results.a4d?projID=EAP314',
+        # 'http://eap.bl.uk/database/results.a4d?projID=EAP458',
         ]
 
 # create a directory to work in and cd into it
@@ -33,7 +34,7 @@ for url in urls:
 
     soup = BeautifulSoup(html_page, 'html.parser')
     results = soup.select('#results')
-    heading = results[0].find('th').h3.text
+    heading = results[0].find('th').h3.text     # heading is the name of the project
     heading = heading.replace(':', '-')
 
     try:
@@ -41,8 +42,12 @@ for url in urls:
     except OSError:
         print "Directory " + heading + " already exists! Not creating it again!"
 
-    with open(os.path.join(abspath+'/'+heading, 'page.html'), 'w') as f:
-        f.write(html_page)
+    # don't write page if already exists
+    if not os.path.exists(os.path.join(abspath+'/'+heading, 'page.html')):
+        with open(os.path.join(abspath+'/'+heading, 'page.html'), 'w') as f:
+            f.write(html_page)
+    else:
+        print "Not writing 'page.html' - already exists"
 
 # list all the directories
 dirs = [d for d in os.listdir('.') if os.path.isdir(d) and not d.startswith('.')]
@@ -51,21 +56,23 @@ for directory in dirs:
     print "Switching to directory: " + directory
     os.chdir(directory)
 
-    # TODO: check if pubs.csv file exists - and write accordingly
+    # check if pubs.csv file exists - and write accordingly
+    if not os.path.exists('pubs.csv'):
+        soup = BeautifulSoup(open('page.html', 'r'), 'html.parser')
+        results = soup.select("#results")
 
-    soup = BeautifulSoup(open('page.html', 'r'), 'html.parser')
-    results = soup.select("#results")
+        base_url = 'http://eap.bl.uk/database/'
 
-    base_url = 'http://eap.bl.uk/database/'
+        with open('pubs.csv', 'wb') as f:
+            writer = csv.writer(f, delimiter='@', quoting=csv.QUOTE_MINIMAL)
+            for tr in results[0].find_all('tr'):
+                for x in range(1, len(tr.find_all('td')), 2):
+                    title = tr.find_all('td')[x].a.text.encode('utf-8').strip()                 # encode to utf-8
+                    link = base_url + tr.find_all('td')[x].a['href'].encode('utf-8').strip()    # encode to utf-8
 
-    with open('pubs.csv', 'wb') as f:
-        writer = csv.writer(f, delimiter='@', quoting=csv.QUOTE_MINIMAL)
-        for tr in results[0].find_all('tr'):
-            for x in range(1, len(tr.find_all('td')), 2):
-                title = tr.find_all('td')[x].a.text.encode('utf-8').strip()                 # encode to utf-8
-                link = base_url + tr.find_all('td')[x].a['href'].encode('utf-8').strip()    # encode to utf-8
-
-                writer.writerow([title, link])
+                    writer.writerow([title, link])
+    else:
+        print "Not writing 'pubs.csv' already exists"
 
     # read the pubs.csv, create a folder for every publication
     with open('pubs.csv', 'r') as f:
@@ -75,17 +82,26 @@ for directory in dirs:
         for row in reader:
             title = row[0].replace('/', '-').replace(':', '-')
             link = row[1]
-            print "trying to load: " + link
-            page = br.open(link).read()
 
-            try:
-                os.mkdir(os.path.join(abspath, title))
-            except OSError:
-                print "Directory: " + title + " already exists! Not creating it again."
+            # skip 1 iteration if zip file of title name exists
+            if os.path.exists(title + '.zip'):
+                print "Already downloaded and zipped " + title
+                continue
 
-            # TODO: don't write if thumbs.html already exists
-            with open(os.path.join(abspath + '/' + title, 'thumbs.html'), 'w') as f:
-                f.write(page)
+            # don't load link and write if thumbs.html already exists
+            if not os.path.exists(os.path.join(title, 'thumbs.html')):
+                print "Loading publication link: " + link
+                page = br.open(link).read()
+
+                try:
+                    os.mkdir(os.path.join(abspath, title))
+                except OSError:
+                    print "Directory: " + title + " already exists! Not creating it again. Not downloading again."
+
+                with open(os.path.join(abspath + '/' + title, 'thumbs.html'), 'w') as f:
+                    f.write(page)
+            else:
+                print "Already exists! folder: " + title + " and thumbs.html"
 
     # get a list of folder names, cd into it, parse the thumbs.html file and store the urls of the image in a file
     folders = [d for d in os.listdir('.') if os.path.isdir(d) and not d.startswith('.')]
@@ -103,26 +119,27 @@ for directory in dirs:
                 image_file_name = image_link.replace('/', '_')
                 image_path = os.path.join(os.path.abspath('.'), image_file_name)
 
-                # download the image
-                # TODO: don't download image if already exists
-                print "Retrieving image: " + full_image_link
-                br.retrieve(full_image_link, image_path)
-                # write the image to file
+                # don't download image if already exists
+                if not os.path.exists(image_file_name):
+                    # download and write the image
+                    print "Retrieving image: " + full_image_link
+                    br.retrieve(full_image_link, image_path)
+                else:
+                    print "Already exists! image: " + image_file_name
+
         except AttributeError:
-            print "AttributeError - There may be no data to parse"
+            print "AttributeError - There may be no images available yet."
             pass
-
-        # filter non-html files and zip it renaming them as 'name_of_current_directory' + '.zip'
-        # check if there are more files than the already thumbs.html
-
-        # TODO: check if a .zip file exists already
-        if len(os.listdir('.')) > 1:
-            for each_file in os.listdir('.'):
-                with zipfile.ZipFile(os.path.basename(os.getcwd()) + ".zip", 'a') as image_zip:
-                    if not each_file.endswith('.html'):
-                        image_zip.write(each_file)
 
         os.chdir('..')
 
+    # zip up all folders
+    for pub_folder in os.listdir('.'):
+        if os.path.isdir(pub_folder):
+            shutil._make_zipfile(pub_folder, pub_folder)    # create a zip archive of the folder
+            shutil.rmtree(pub_folder)                       # delete the folder & contents
+
     # return to previous directory
     os.chdir('..')
+
+print "DONE!"
